fix: 🐛 create config directory if it does not exist
Adding a series to watching list will fail if config directory does not
exist. Before this fix, the script checks if the config directory exists
only when user choose to create a custom config file for a series. Now,
the script will always create a config directory if it does not exist
when it runs.