fix: 🐛 handle parse playlist-index error
If an episode's format does not match the specified filter when
playlist-index is parsed, youtube-dl (with --ignore-errors flag) throws
an error message: 'requested format not found'. Then it continues
parsing the remaining episodes in the playlist. However, it returns
error code 1 when completed. In this case, the `if` condtional statement
always evaluates to `else` expression block. This fix ignores the code 1
errors and make sure that the generated cached playlist-index file is
not empty to evaluate to a successful operation.