feat: 🎸 preview selected confing file
When user choose to select a config file for a series, user loose the
ability to preview it if only one config file was found -- because it
will be selected by default. This feature allow the user to see a
preview of the file when user is prompted to customize it.