fix: 🐛 ensure that only fragmented files will be deleted
In instance when fragment monitoring process starts reading previous
download log before it was overwritten by new download process (which
was fixed already) and after the download process is terminated, the
fragmented file list will be empty. Because of that, find command will
list all `mp4` files in that directory for deletion. Now, a conditional
statement was added to kill the download process in such instance and
safeguard user's files.