feat: 🎸 obtain user config based on youtube-dl extractor
User config for an extracotr is obtained automatically if it is
available in `clanime` config directory or user specified directory
(when it was setup as an environment variable: `CLANIME_EXTRACTORS_CONFIG_DIR`).
Extractor config filename must exactly matches youtube-dl extractor name
(or the first part of it if the name has two portions delimited by a colon).
Because extractor name is parsed along series title via youtube-dl (if series
is not saved in a list), user must pass the extractor name with `-e` or `--extractor`
option if authentication details (e.g. cookies and user-agent) that are
required for that extractor is saved in extractor config file.