feat: 🎸 override download-archive path with cli argument
Allow user to override download-archive file path with youtube-dl
`--download-arcive` `<PATH>` arguments. Now, the behavior of clanime
maches the expected behavior of `youtube-dl` when this option is used --
when a filename is provided without a path, the current working
directory (where clanime was excuted) will be used as the base directory
for archive file. File full path will be used when provided by user.