feat: 🎸 prompt user to try download again if Error 403 occured
HTTP error 403 might occur due to invalid credentials when youtube-dl
download process starts. User will be given the choice to try the
download again without terminating the process. This is useful when, for
instance, cookies file is used for authentication. A tip for providing
a new cookies file will be displayed if cookies option was found in
extractor config file or cli arguments.